+++ src/client/routes/_layout.create.$puzzleId.tsx
@@ -20,7 +20,10 @@ export const Route = createFileRoute('/_layout/create/$puzzleId')({
     } catch (error) {
       toast.error((error as Error).message);
       throw redirect({
-        to: '/create',
+        to: '/solve',
+        params: {
+          puzzleId: params.puzzleId,
+        },
       });
     }
   },